move run functions from lib to harness/run.rs
