+++ src/bin/run.rs
@@ -12,21 +12,21 @@ pub fn main(
     log_handler: MultiProgress,
 ) -> Result<()> {
     let experiment = experiment.canonicalize()?;
-    if experiment.display().to_string() == "." {
+    if experiment == std::env::current_dir()? {
         return Err(Error::HarnessRunError {
-            experiment: file_name_string(&experiment),
-            err: "Cannot start experiment run from the experiment source folder.".to_string(),
+            experiment: file_name_string(&experiment.canonicalize()?),
+            err: "Cannot start experiment run from pwd.".to_string(),
         });
     }
 
     if let Some(env) = trial {
-        exomat::run_trial(&experiment, env, log_handler)
+        exomat::harness::run::trial(&experiment, env, log_handler)
     } else {
         let output = match output {
             Some(x) => Ok(x),
             None => exomat::harness::skeleton::generate_build_series_filepath(&experiment),
         }?;
 
-        exomat::run_experiment(&experiment, repetitions, output, log_handler)
+        exomat::harness::run::experiment(&experiment, repetitions, output, log_handler)
     }
 }
